+++ Source/ProgramRootVC.m
@@ -167,7 +167,7 @@
     self.dayNext.enabled = dayIndex != self.dayController.count - 1;
     
     ProgramDayTVC *pdvc = self.dayController[dayIndex];
-    self.dayLabel.title = [NSString stringWithFormat:@"%0.2d.%0.2d", pdvc.day.date.day, pdvc.day.date.month];
+    self.dayLabel.title = [NSString stringWithFormat:@"%0.2ld.%0.2ld", (long)pdvc.day.date.day, (long)pdvc.day.date.month];
     
     _dayIndex = dayIndex;
 }
